+++ sdk/autoload.php
@@ -4,4 +4,4 @@
 
 require_once __DIR__ . '/composer/autoload_real.php';
 
-return ComposerAutoloaderInit21e7c608f8f183f08e6d87d77b1821bf::getLoader();
+return ComposerAutoloaderInit784958e6477c40b8c05900cbaf08eda2::getLoader();